wxo-talent-acquisition/pyproject.toml
[project]
name = "texas-state-bank-talent-acquisition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.29.5",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "appnope==0.1.4",
    "asttokens==3.0.0",
    "cachetools==6.1.0",
    "certifi==2025.7.9",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "comm==0.2.2",
    "debugpy==1.8.14",
    "decorator==5.2.1",
    "elastic-transport==8.17.1",
    "elasticsearch==9.0.2",
    "executing==2.2.0",
    "fastapi==0.116.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "ibm-cos-sdk==2.14.2",
    "ibm-cos-sdk-core==2.14.2",
    "ibm-cos-sdk-s3transfer==2.14.2",
    "ibm-watsonx-ai==1.3.28",
    "idna==3.10",
    "ipykernel==6.29.5",
    "ipython==9.4.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jmespath==1.0.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "lomond==0.3.3",
    "matplotlib-inline==0.1.7",
    "nest-asyncio==1.6.0",
    "numpy==2.3.1",
    "packaging==25.0",
    "pandas==2.2.3",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "platformdirs==4.3.8",
    "prompt-toolkit==3.0.51",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "pyzmq==27.0.0",
    "requests==2.32.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "stack-data==0.6.3",
    "starlette==0.46.2",
    "tabulate==0.9.0",
    "tornado==6.5.1",
    "traitlets==5.14.3",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "wcwidth==0.2.13",
    "websockets==15.0.1",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.20",
]

